[Code together] Make QFT a template instead of an operation issue (#1548)
* changes for issue 1477

* Update CHANGELOG.md

* formatted by black

* update test_qft

* Update test_qft.py

* Update test_qft.py

* Update test_qft.py

* disable ops qft & its tests

* disable ops qft

* Update non_parametric_ops.py

* fix qft

* fix qft

* Update qmc.py

* Update test_qubit_ops.py

* Update pennylane/ops/qubit/non_parametric_ops.py


* Update pennylane/ops/qubit/non_parametric_ops.py


* Update tests/templates/test_subroutines/test_qft.py


* Update non_parametric_ops.py

* change the import

* change qml.QFT to qml.templates.QFT

* Update operations.rst

* Update non_parametric_ops.py

* Update non_parametric_ops.py

* Update CHANGELOG.md

* Update CHANGELOG.md
