+++ pennylane/templates/subroutines/__init__.py
@@ -27,3 +27,4 @@ from .qpe import QuantumPhaseEstimation
 from .qmc import QuantumMonteCarlo
 from .all_singles_doubles import AllSinglesDoubles
 from .grover import GroverOperator
+from .qft import QFT
