+++ pennylane/transforms/qmc.py
@@ -15,9 +15,10 @@
 Contains the quantum_monte_carlo transform.
 """
 from functools import wraps
-from pennylane import PauliX, Hadamard, MultiControlledX, CZ, QFT
+from pennylane import PauliX, Hadamard, MultiControlledX, CZ
 from pennylane.wires import Wires
 from pennylane.transforms import adjoint
+from pennylane.templates import QFT
 
 
 def _apply_controlled_z(wires, control_wire, work_wires):
